[CSL-987] Get rid of SafeCopy in most places
