--- src/Pos/SafeCopy/GodTossing.hs
@@ -1,5 +1,0 @@
--- | GodTossing serialization as SafeCopy.
-
-{-# OPTIONS_GHC -F -pgmF autoexporter #-}
-{-# OPTIONS_GHC -Wno-unused-imports   #-}
-{-# OPTIONS_GHC -Wno-dodgy-exports    #-}
