--- src/Pos/SafeCopy/GodTossing/Types.hs
@@ -1,13 +1,0 @@
-
--- | SafeCopy serialization of types from GodTossing.Types.Types.
-
-module Pos.SafeCopy.GodTossing.Types
-       (
-       ) where
-
--- import           Data.SafeCopy                  (base, deriveSafeCopySimple)
-
--- import           Pos.SafeCopy.GodTossing.Base   ()
--- import           Pos.Ssc.GodTossing.Types.Types (GtPayload (..), GtProof (..))
-
--- deriveSafeCopySimple 0 'base ''GtProof
