--- src/Pos/SafeCopy/Update.hs
@@ -1,12 +1,0 @@
-
--- | SafeCopy serialization of core Update System types.
-
-module Pos.SafeCopy.Update
-       (
-       ) where
-
--- import           Data.SafeCopy         (base, deriveSafeCopySimple)
-
--- FIXME
--- import           Pos.Core.Types
--- import           Pos.Update.Core.Types
